Fixed A fatal flaw that caused the roborio to shutdown
Fixed A fatal flaw that caused the roborio to shutdown. When running the
I2C version fo the navX code and theSF2 software a file was crossed
linked which caused the roborio to stop working. I fixed this issue and
the fault went away. You can tell if the fault is still there because it
throws a fault on the driver station error window. The fault says
something about a wrong path or invalid charater.